src/requirements.txt
aiofiles==0.5.0
aniso8601==7.0.0
async-exit-stack==1.0.1
async-generator==1.10
bcrypt==3.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
click==7.1.2
dnspython==2.0.0
ecdsa==0.14.1
email-validator==1.1.2
fastapi==0.62.0
graphene==2.1.8
graphql-core==2.3.2
graphql-relay==2.0.1
h11==0.9.0
httptools==0.1.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
motor==2.3.0
orjson==3.4.5
passlib==1.7.4
promise==2.3
pyasn1==0.4.8
pycparser==2.20
pydantic==1.7.3
pymongo==3.11.2
python-jose==3.2.0
python-multipart==0.0.5
PyYAML==5.3.1
requests==2.25.0
rsa==4.6
Rx==1.6.1
six==1.15.0
starlette==0.13.6
ujson==3.2.0
urllib3==1.26.2
uvicorn==0.11.8
uvloop==0.14.0
websockets==8.1
python-decouple==3.3
